requirements.txt
matplotlib
python-dotenv
watchdog
requests